requirements.txt
django~=3.2.0
pillow
djangorestframework
django-cors-headers
black
djangorestframework-simplejwt